Removed timeout from check_consistency() test
It's flaky with timeout. The reason timeout was added is that if there
is an error in implementation test might hang. But I beleive it's a
better to hang when error present (it help ti get diagnostics) than to
fail periodically on timeout even if there is nothing wrong with
implementation.